Refactor: libcrmservice: separate LSB-specific code into own source files
... to be more orthogonal to systemd and upstart. No code changes other
than some function renames.
