+++ lib/services/Makefile.am
@@ -1,19 +1,8 @@
-# Copyright (c) 2012 David Vossel <davidvossel@gmail.com>
 #
-# This library is free software; you can redistribute it and/or
-# modify it under the terms of the GNU Lesser General Public
-# License as published by the Free Software Foundation; either
-# version 2.1 of the License, or (at your option) any later version.
-# 
-# This library is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-# Lesser General Public License for more details.
-# 
-# You should have received a copy of the GNU Lesser General Public
-# License along with this library; if not, write to the Free Software
-# Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
+# Copyright 2012-2018 David Vossel <davidvossel@gmail.com>
 #
+# This source code is licensed under the GNU Lesser General Public License
+# version 2.1 or later (LGPLv2.1+) WITHOUT ANY WARRANTY.
 #
 
 MAINTAINERCLEANFILES = Makefile.in
@@ -21,7 +10,8 @@ MAINTAINERCLEANFILES = Makefile.in
 AM_CPPFLAGS			= -I$(top_srcdir)/include
 
 lib_LTLIBRARIES			= libcrmservice.la
-noinst_HEADERS			= pcmk-dbus.h upstart.h systemd.h services_private.h
+noinst_HEADERS			= pcmk-dbus.h upstart.h systemd.h \
+				  services_lsb.h services_private.h
 
 libcrmservice_la_LDFLAGS	= -version-info 8:1:0
 libcrmservice_la_CPPFLAGS	= -DOCF_ROOT_DIR=\"@OCF_ROOT_DIR@\" $(AM_CPPFLAGS)
@@ -32,7 +22,7 @@ libcrmservice_la_LDFLAGS	+= $(LDFLAGS_HARDENED_LIB)
 
 libcrmservice_la_LIBADD		= $(GIO_LIBS) $(top_builddir)/lib/common/libcrmcommon.la $(DBUS_LIBS)
 
-libcrmservice_la_SOURCES	= services.c services_linux.c
+libcrmservice_la_SOURCES	= services.c services_linux.c services_lsb.c
 if BUILD_DBUS
 libcrmservice_la_SOURCES	+= dbus.c
 endif
