+++ lib/services/services_private.h
@@ -1,19 +1,8 @@
 /*
- * Copyright (C) 2010 - 2011, Red Hat, Inc.
+ * Copyright 2010-2018 Red Hat, Inc.
  *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 2.1 of the License, or (at your option) any later version.
- *
- * This software is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ * This source code is licensed under the GNU Lesser General Public License
+ * version 2.1 or later (LGPLv2.1+) WITHOUT ANY WARRANTY.
  */
 
 #ifndef SERVICES_PRIVATE__H
@@ -53,9 +42,6 @@ GList *services_os_get_directory_list(const char *root, gboolean files, gboolean
 
 G_GNUC_INTERNAL
 gboolean services_os_action_execute(svc_action_t * op);
-
-G_GNUC_INTERNAL
-GList *resources_os_list_lsb_agents(void);
 
 G_GNUC_INTERNAL
 GList *resources_os_list_ocf_providers(void);
